Move osrs.py to the same folder as other functions
