+++ PyChan/Core/Commands/Games/Functions/osrs.py
