+++ PyChan/Core/Commands/Games/games.py
@@ -1,6 +1,6 @@
 from nextcord.ext import commands
 
-from Core.Commands.Games.Functions.Osrs.osrs import Osrs
+from Core.Commands.Games.Functions.osrs import Osrs
 from Core.Commands.Games.Functions.osu import Osu
 from Core.Commands.Games.Functions.summoner_info import SummonerInfo
 from Core.Commands.Games.Functions.last_match import LastMatch
